requirements.txt
numpy
Pillow
pkg-resources==0.0.0
six==1.12.0
tensorboardX==1.8
torch==1.7.1
torchvision==0.4.0
tqdm==4.35.0
opencv-python==4.4.0.46
imageio==2.9.0
streamlit